Started again after a long hiatus!
I taken a long break from development in the middle of creating a system for object persistence between rooms. I've come back to a mess of code which I've mostly forgotten. Objects weren't actually warping to other rooms, which I've corrected. There are still many problems though, starting with the fact that I can't carry objects to the other room