Re-doing the persistence system
It was a mess. I'm going with a new gameplay and starting to rebuild it. Now, only the player character will be able to traverse scenes. He will be able to carry objects with him, but any objects thrown at a warp trigger will bounce off.

Still in the early stages of rebuilding. I have a system setup to warp the character, but it appears to not be working. I made the character persistant, which is a change from how I was doing it previously. We can not have only one player. Have not started interation between warp trigger and other objects / carried objects