FINALLY Fixed warping and icons!
FINALLY have this stuff fixed. I improved the warping system so we now relocate the character in MapSystem instead of WarpTrigger. Fixed the problem with the action icons that occured when we child icons for scene change. Objects can be carried. More testing and improving is probably needed, but I'm FINALLY about ready to move onto the next thing.