Implemented MapSystem::InstantiateOnLoad, fixed some scene fade transition issues
Added a system for instantiating objects on next scene load. Could be fleshed out more. This addition (which is going to be needed anyway) was added in reponse to a horrible glitch in the scene fade transition that took forever to track down. Turns out you can't instantiate objects immediately after the scene load function is called. You have to use a callback to know when the scene is fully loaded first.