Fixed the aim reticle! Also fixed some camera/background glitches
The pickup and aim action feature now seems to work as intended! It's about time. I had a horrible glitch that took awhile to fix. It appears to work correctly now! WOO!

There was also a glitch introduced with the background when I changed some of the camera code a few updates ago. Made some changes to background and cameraFollowPlayer to fix it.