Graphic style got a nice update.
I think this is the art style I'm going with. The camera seems to jitter some when it is at the scene boundary, and I haven't figured out how to fix it. Have a few glitches to fix with the conveyor and some other mild glitches. Need to add oxygen meter in water, missile, doors, possibly a cabinet and inventory system. Dress system needs updated. Then sound needs added. Still a lot to do before level design can really begin!