Improved character facing system, updated dress and pickup to reflect. Added dresser!
Not sure when the last update was, but I've done a lot of different stuff. Dresses and pickup items can now face different directions with the character, though pickupobject.m_carrytype.exactPlayerPosition now needs to be setup to follow player exactly.

Added a dresser! As well as a range script that triggers events specified in the editor. Some stuff to fix and adjust, but this is some nice progress.