Refined react, changed breakOnImpact to impactThreshold, added requireParentTags to actionInRange
Added some stuff to react. Improved the way dialog interupt is handled. Added a variable to forward interupts. Added Ouch. Added faceInitiator.

Changed breakOnImpact to impactThreshold. It's now more multipurpose and can be used for things like sending an ouch command when you hit someone in the face with a ball.

Added requireParentTags to actionInRange to prevent Pinky from showing the pickupObject range icon. We are now restricting range interaction to the player.