FIXED THE CAMERA ISSUE! Added gmx level importer
I fixed the issue with the camera not clamping properly. I am now using a different way to calculate viewport width and height, which corrected the problem. Added drawClampLines to cameraFollowPlayer and drawBoundaries to sceneBoundaries for debug purposes.

Also got a basic GMX level importer working! We have the start of the first level of Blubber one! This will make it easy to import levels from the first game once it's complete.