Added AI_KeepObjectsPickedUp
Started the vacuum cleaner as well. I'm gonna have to take a pause on the cpuinput script and work on other stuff, because I've put entirely too much time into it. It's mostly functional it seems. AI_KeepObjectsPickedUp could definitely use some more work. There seems to be a bit of glitching that happens with object pickup. The bare bones are there though.